Make it clearer bosh metrics are not default
Making this change because, as written, people will think they get bosh health metrics in firehose by default, that's not clear unless you keep reading a few paragraphs. This is something people already get confused on all the time. We can't get them into the firehose by default until PCF 1.12-1.13. I also eliminated the word "logs" in "logs and metrics" because we never talk about logs anywhere else on this page, and by saying "logs" you'd have to get into all the ways logs actually can/do get into the firehose for transmission as well as the log paths outside the firehose. Leaving it in could worsen confusion we already deal with that the firehose is not the only 'of interest to operators/devs' log transport mechanism in the platform.