Update index.html.md.erb
The nsync bulker is essentially gone, and so is this nsync_bulker recommendation now. Removing outright for 1.11 after convo with CC team (and Diego team). Feel free to copy edit the reason. It was hard for Zach and I to come up with a good customer-facing version of the reason for deletion (vs replacement)