Update kpi.html.md.erb
As written this is confusing a lot of people. Contextually it should work similarly to the Remaining Disk warning, which is not confusing people. The calculation stated was very specific to DD manipulation of how the data appears graphed. I’m updating this to better align with the less confusing Minimum Disk recommendation, which then makes more sense with the actual threshold values stated, and should be a clearer interpretation resulting in the intended purpose for the alert recommendation