Add known issue to metric for v2.0
PCF v2.0 is not currently emitting this metric (cf is). Background https://pivotal.slack.com/archives/C0555LE4W/p1520965583000803. Being resolved in https://www.pivotaltracker.com/story/show/155958008. Best approach is to add a known issue in the meantime, and then update it with the given PAS version fix when this is finished being backpatched.