Remove incorrect PAS MySQL recommendations
The removed recommendations are not valid. This was previously noted and expected to be corrected in https://www.pivotaltracker.com/n/projects/923920/stories/157683592. Because this work has still not yet occurred, I am removing these invalid recommendations until the corrections can be made, as we do not want customers to invest effort to monitor these as they are currently documented.